pyproject.toml
[tool.poetry]
name = "incident-tracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.121.0"
sqlalchemy = "^2.0.44"
uvicorn = "^0.38.0"
gunicorn = "^23.0.0"
pytz = "^2025.2"
alembic = "^1.17.1"
pydantic-settings = "^2.11.0"
pyjwt = "^2.10.1"
redis = "^7.0.1"
cryptography = "^46.0.3"
asyncpg = "^0.30.0"
pre-commit = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]

force-exclude = true
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
extend-exclude = ["src/infrastruction/db/migrations/env.py"]

line-length = 100
indent-width = 4
target-version = "py312"

[tool.ruff.lint]

select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]

quote-style = "double"
indent-style = "tab"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint.isort]
case-sensitive = true
lines-between-types = 1
section-order = [
  "future",
  "standard-library",
  "third-party",
  "first-party",
  "local-folder",
]
